pyproject.toml
[tool.poetry]
name = "госвнииас"
version = "0.1.0"
description = ""
authors = ["DEDMOPO3PEAHIMATOP <ssp2002@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.0.0"
torch = "^2.0.1"
torchvision = "^0.15.2"
ipython = "^8.15.0"
matplotlib = "^3.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
